+++ CHANGELOG.md
@@ -8,10 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 * Can now connect to local endpoints via TLS
+* wh-server forwards ingress traffic from a single shared TLS port via SNI
 
 ### Fixed
 * Race condition with session access in remote/http2 (#26)
 * Errant `FLY_ENDPOINT` references in usage output
+
 
 
 ## [0.5.36] - 2017-10-09